+++ calypso_anemometer/core.py
@@ -8,52 +8,87 @@ References
 - https://github.com/hbldh/bleak/blob/develop/examples/get_services.py
 - https://github.com/hbldh/bleak/blob/develop/examples/service_explorer.py
 """
-import asyncio
+import concurrent
 import logging
-import os
-import sys
 import time
+from typing import Optional
 
-from bleak import BleakClient, BleakScanner
+from bleak import BleakClient, BleakError, BleakScanner
+
+from calypso_anemometer.exception import BluetoothAdapterError, BluetoothConversationError
 
 # Configuration section.
 DISCOVERY_TIMEOUT = 7.5
 DEVICE_TIMEOUT = 10.0
 BLUETOOTH_ADAPTER = "hci0"
-DEVICE_ADDRESS = os.getenv("CALYPSO_ADDRESS")
 
 logger = logging.getLogger()
 
 
-async def find_device_address():
-    """
-    Find Calypso device.
-    TODO: Use https://github.com/hbldh/bleak/blob/develop/examples/scanner_byname.py.
-    """
-    if DEVICE_ADDRESS is not None:
-        return DEVICE_ADDRESS
-    logger.info("Using BLE discovery to find Calypso ULTRASONIC anemometer")
-    devices = await BleakScanner.discover(timeout=DISCOVERY_TIMEOUT, adapter=BLUETOOTH_ADAPTER)
-    for device in devices:
-        if device.name == "ULTRASONIC":
-            logger.info(f"Found device: {device}")
-            # Add some delay between discovery and readout. It looks like this stabilizes readings.
-            time.sleep(0.5)
-            return device.address
-    logger.error("Unable to find device")
+class CalypsoDeviceApi:
 
+    NAME = "calypso-up10"
+    DESCRIPTION = "Calypso UP10 anemometer"
 
-def get_adapter_name(client):
-    if hasattr(client, "_adapter"):
-        return client._adapter
+    def __init__(self, ble_address: Optional[str] = None):
+        self.ble_address = ble_address
+        self.client: BleakClient = None
 
+    async def discover(self, force=False) -> bool:
+        """
+        Discover device via BLE.
+        TODO: Use https://github.com/hbldh/bleak/blob/develop/examples/scanner_byname.py.
+        """
+        # Skip discovery when already discovered and not forced.
+        if self.ble_address is not None and not force:
+            return True
 
-async def run_conversation(device_address):
-    client = BleakClient(device_address, timeout=DEVICE_TIMEOUT, adapter=BLUETOOTH_ADAPTER)
-    logger.info(f"Connecting to device at {device_address} with {get_adapter_name(client)}")
-    try:
-        await client.connect()
-        services = await client.get_services()
+        logger.info(f"Using BLE discovery to find {self.DESCRIPTION}")
+        devices = await BleakScanner.discover(timeout=DISCOVERY_TIMEOUT, adapter=BLUETOOTH_ADAPTER)
+        for device in devices:
+            if device.name == "ULTRASONIC":
+                logger.info(f"Found device at address: {device}")
+                # Add some delay between discovery and readout. It looks like this stabilizes readings.
+                time.sleep(0.5)
+                self.ble_address = device.address
+                return True
+        logger.error("Unable to find device")
+        return False
+
+    async def connect(self):
+        self.client = BleakClient(self.ble_address, timeout=DEVICE_TIMEOUT, adapter=BLUETOOTH_ADAPTER)
+        logger.info(f"Connecting to device at {self.ble_address} with adapter {get_adapter_name(self.client)}")
+        try:
+            await self.client.connect()
+        except (BleakError, concurrent.futures.TimeoutError) as ex:
+            message = f"{ex.__class__.__name__}: {ex}"
+            logger.error(f"Conversation went south: {message}")
+            if "Bluetooth device is turned off" in message:
+                raise BluetoothAdapterError(message) from None
+            else:
+                raise BluetoothConversationError(message) from None
+        # finally:
+        #    logger.info("Disconnecting")
+        #    await self.disconnect()
+
+    async def disconnect(self):
+        logger.info("Disconnecting")
+        try:
+            await self.client.disconnect()
+        except Exception:
+            logger.exception("Disconnect failed")
+
+    async def explore(self):
+        """
+        Explore all services and characteristics. Useful for debugging purposes.
+
+        Possible errors:
+
+        - bleak.exc.BleakDBusError: [org.bluez.Error.Failed] Software caused connection abort
+        - bleak.exc.BleakError: Device with address E7:B6:1B:DB:02:DF was not found.
+        - futures.TimeoutError:
+        """
+        services = await self.client.get_services()
         for service in services:
             logger.info(f"Found service: {service}")
             for char in service.characteristics:
@@ -62,7 +97,7 @@ async def run_conversation(device_address):
                 value = None
                 if "read" in char.properties:
                     try:
-                        value = bytes(await client.read_gatt_char(char.uuid))
+                        value = bytes(await self.client.read_gatt_char(char.uuid))
                     except Exception as e:
                         logger.exception(f"  Reading characteristic failed: {char}")
                 logger.info(f"  Value: {value}")
@@ -71,32 +106,18 @@ async def run_conversation(device_address):
                     logger.info(f"    Found descriptor: {descriptor}")
                     value = None
                     try:
-                        value = bytes(await client.read_gatt_descriptor(descriptor.handle))
+                        value = bytes(await self.client.read_gatt_descriptor(descriptor.handle))
                     except Exception as e:
                         logger.exception(f"    Reading descriptor failed: {descriptor}")
                     logger.info(f"    Value: {value}")
 
-    except Exception as ex:
-        # bleak.exc.BleakDBusError: [org.bluez.Error.Failed] Software caused connection abort
-        # bleak.exc.BleakError: Device with address E7:B6:1B:DB:02:DF was not found.
-        # futures.TimeoutError:
-        logger.exception("Conversation went south")
-    finally:
-        logger.info("Disconnecting")
-        await client.disconnect()
+    async def get_info(self):
+        ch = await self.client.read_gatt_char("00002a29-0000-1000-8000-00805f9b34fb")
+        print(ch.decode())
+        # print(dir(ch))
+        # print(ch.V)
 
 
-def setup_logging(level=logging.INFO):
-    log_format = "%(asctime)-15s [%(name)-8s] %(levelname)-8s: %(message)s"
-    logging.basicConfig(format=log_format, stream=sys.stderr, level=level)
-
-
-async def main():
-    setup_logging(level=logging.INFO)
-    device_address = await find_device_address()
-    if device_address is None:
-        sys.exit(1)
-    await run_conversation(device_address)
-
-
-asyncio.run(main())
+def get_adapter_name(client):
+    if hasattr(client, "_adapter"):
+        return client._adapter