+++ calypso_anemometer/exception.py
@@ -1,0 +1,6 @@
+class BluetoothAdapterError(Exception):
+    pass
+
+
+class BluetoothConversationError(Exception):
+    pass