+++ CHANGES.rst
@@ -6,5 +6,9 @@ calypso-anemometer changelog
 in progress
 ===========
 
+- Minimal implementation, connecting to the device
+- Add CLI interface with subcommands ``info`` and ``explore``
+
+
 2022-xx-xx 0.0.0
 ================